requirements.txt
Django>=2.1.1, <3.0.0
django-crispy-forms==1.7.2
Pillow==5.4.1